스타일링: "배너 컴포넌트 추가"
- 배너 컴포넌트 추가
- 배너 이미지 추가
