+++ src/app/layout.tsx
@@ -1,12 +1,10 @@
 import type { Metadata } from 'next';
-import { Noto_Sans_KR, Ubuntu } from 'next/font/google';
+import { Noto_Sans_KR } from 'next/font/google';
 import './globals.css';
 import { NextLayout, NextProvider } from './provider';
 import { cn } from '@/lib/utils';
 
 const noto = Noto_Sans_KR({ subsets: ['latin'] });
-
-const ubuntu = Ubuntu({ weight: ['400', '500', '700'], subsets: ['latin'] });
 
 export const metadata: Metadata = {
   title: 'Create Next App',
@@ -20,7 +18,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="ko">
-      <body className={cn(noto.className, ubuntu.className)}>
+      <body className={cn(noto.className)}>
         <NextProvider>
           <NextLayout>{children}</NextLayout>
         </NextProvider>
